server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.9
python-dotenv==1.1.1
google-cloud-aiplatform==1.113.0
google-cloud-firestore==2.21.0
